kvc-migrate-init-pools: Migrate _init_pools from mixin to KVCacheConfigurator
Refactor chain ID: kvc-migrate-init-pools
