+++ python/sglang/srt/model_executor/model_runner_kv_cache_mixin.py
@@ -5,48 +5,12 @@ from typing import TYPE_CHECKING
 
 import torch
 
-from sglang.srt.configs.hybrid_arch import (
-    hybrid_gdn_config,
-    mambaish_config,
-)
 from sglang.srt.configs.model_config import (
-    get_dsa_index_head_dim,
-    is_deepseek_dsa,
     is_deepseek_v4,
 )
 from sglang.srt.environ import envs
-from sglang.srt.layers.dp_attention import get_attention_tp_size
-from sglang.srt.mem_cache.allocator import (
-    PagedTokenToKVPoolAllocator,
-    TokenToKVPoolAllocator,
-)
-from sglang.srt.mem_cache.allocator.hisparse import (
-    DeepSeekV4HiSparseTokenToKVPoolAllocator,
-    HiSparseTokenToKVPoolAllocator,
-)
-from sglang.srt.mem_cache.allocator.swa import SWATokenToKVPoolAllocator
-from sglang.srt.mem_cache.common import get_req_to_token_extra_context_len
-from sglang.srt.mem_cache.deepseek_v4_memory_pool import DeepSeekV4TokenToKVPool
-from sglang.srt.mem_cache.hisparse_memory_pool import HiSparseDSATokenToKVPool
-from sglang.srt.mem_cache.kv_cache_configurator import (
-    calculate_mla_kv_cache_dim,
-)
-from sglang.srt.mem_cache.memory_pool import (
-    DSATokenToKVPool,
-    HybridLinearKVPool,
-    HybridReqToTokenPool,
-    MHATokenToKVPool,
-    MHATokenToKVPoolFP4,
-    MLATokenToKVPool,
-    MLATokenToKVPoolFP4,
-    NoOpMHATokenToKVPool,
-    ReqToTokenPool,
-)
-from sglang.srt.mem_cache.swa_memory_pool import SWAKVPool
-from sglang.srt.platforms import current_platform
 from sglang.srt.utils.common import (
     get_available_gpu_memory,
-    is_float4_e2m1fn_x2,
     is_hip,
     is_npu,
 )
@@ -97,675 +61,6 @@ class ModelRunnerKVCacheMixin:
     def _calculate_mamba_ratio(self: ModelRunner) -> int:
         return self.kv_cache_configurator._calculate_mamba_ratio()
 
-    def _validate_prefill_only_disable_kv_cache_pool_family(
-        self: ModelRunner,
-        is_dsa_model: bool,
-        is_dsv4_model: bool,
-        current_platform,
-    ):
-        if not self.server_args.prefill_only_disable_kv_cache or self.is_draft_worker:
-            return
-
-        unsupported_pool_family = None
-        if is_dsv4_model:
-            unsupported_pool_family = "DeepSeekV4TokenToKVPool"
-        elif current_platform.is_out_of_tree() and not mambaish_config(
-            self.model_config
-        ):
-            unsupported_pool_family = "out-of-tree platform KV pool"
-        elif self.server_args.attention_backend == "ascend" and not mambaish_config(
-            self.model_config
-        ):
-            unsupported_pool_family = "NPU/Ascend KV pool"
-        elif self.use_mla_backend and is_dsa_model:
-            unsupported_pool_family = "DSA/MLA KV pool"
-        elif self.use_mla_backend and not mambaish_config(self.model_config):
-            unsupported_pool_family = "MLA KV pool"
-        elif self.is_hybrid_swa:
-            unsupported_pool_family = "SWA KV pool"
-        elif mambaish_config(self.model_config):
-            unsupported_pool_family = "hybrid linear/Mamba KV pool"
-        elif is_float4_e2m1fn_x2(self.kv_cache_dtype):
-            unsupported_pool_family = "FP4 MHA KV pool"
-
-        if unsupported_pool_family is not None:
-            raise RuntimeError(
-                "--prefill-only-disable-kv-cache is not supported for "
-                f"{unsupported_pool_family}. Supported configurations today: plain MHA "
-                "models on CUDA with the FA (fa3/fa4) prefill backend, --is-embedding, "
-                "--chunked-prefill-size=-1, --disable-radix-cache, no context-parallel "
-                "attention, no HiSparse, and --kv-cache-dtype != fp4_e2m1."
-            )
-
-    def _init_pools(self: ModelRunner):
-        """Initialize the memory pools."""
-        max_num_reqs = self.max_running_requests
-
-        # Initialize req_to_token_pool
-        if self.req_to_token_pool is None:
-            max_spec_draft_tokens = self.server_args.max_speculative_num_draft_tokens
-            extra_max_context_len = get_req_to_token_extra_context_len(self.server_args)
-
-            if self.server_args.disaggregation_mode == "decode":
-                from sglang.srt.disaggregation.decode import (
-                    DecodeReqToTokenPool,
-                    HybridMambaDecodeReqToTokenPool,
-                )
-
-                # Extra slots for pre-allocated requests
-                pre_alloc_size = self.server_args.disaggregation_decode_extra_slots
-                if config := mambaish_config(self.model_config):
-                    self.req_to_token_pool = HybridMambaDecodeReqToTokenPool(
-                        size=max_num_reqs,
-                        max_context_len=self.model_config.context_len
-                        + extra_max_context_len,
-                        device=self.device,
-                        enable_memory_saver=self.server_args.enable_memory_saver,
-                        cache_params=config.mamba2_cache_params,
-                        mamba_layer_ids=(
-                            [
-                                i
-                                for i in config.mamba2_cache_params.layers
-                                if self.start_layer <= i < self.end_layer
-                            ]
-                        ),
-                        speculative_num_draft_tokens=max_spec_draft_tokens,
-                        speculative_eagle_topk=self.server_args.speculative_eagle_topk,
-                        enable_mamba_extra_buffer=self.server_args.enable_mamba_extra_buffer(),
-                        pre_alloc_size=pre_alloc_size,
-                        enable_overlap_schedule=not self.server_args.disable_overlap_schedule,
-                        mamba_size=self.server_args.max_mamba_cache_size,
-                        start_layer=self.start_layer,
-                    )
-                else:
-                    self.req_to_token_pool = DecodeReqToTokenPool(
-                        size=max_num_reqs,
-                        max_context_len=self.model_config.context_len
-                        + extra_max_context_len,
-                        device=self.device,
-                        enable_memory_saver=self.server_args.enable_memory_saver,
-                        pre_alloc_size=pre_alloc_size,
-                    )
-            elif config := mambaish_config(self.model_config):
-                self.req_to_token_pool = HybridReqToTokenPool(
-                    size=max_num_reqs,
-                    mamba_size=self.server_args.max_mamba_cache_size,
-                    mamba_spec_state_size=max_num_reqs,
-                    max_context_len=self.model_config.context_len
-                    + extra_max_context_len,
-                    device=self.device,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    cache_params=config.mamba2_cache_params,
-                    mamba_layer_ids=(
-                        [
-                            i
-                            for i in config.mamba2_cache_params.layers
-                            if self.start_layer <= i < self.end_layer
-                        ]
-                    ),
-                    enable_mamba_extra_buffer=self.server_args.enable_mamba_extra_buffer(),
-                    enable_mamba_extra_buffer_lazy=self.server_args.enable_mamba_extra_buffer_lazy(),
-                    speculative_num_draft_tokens=max_spec_draft_tokens,
-                    speculative_eagle_topk=self.server_args.speculative_eagle_topk,
-                    enable_overlap_schedule=not self.server_args.disable_overlap_schedule,
-                    start_layer=self.start_layer,
-                )
-            else:
-                # DSV4 on NPU needs an extended ReqToTokenPool holding per-req
-                # swa/c4/c128/c{4,128}_state tables; others stay on the stock one.
-                req_to_token_pool_cls = ReqToTokenPool
-                if _is_npu and is_deepseek_v4(self.model_config.hf_config):
-                    from sglang.srt.hardware_backend.npu.dsv4.dsv4_req_to_token_pool import (
-                        DSV4NPUReqToTokenPool,
-                    )
-
-                    req_to_token_pool_cls = DSV4NPUReqToTokenPool
-
-                self.req_to_token_pool = req_to_token_pool_cls(
-                    size=max_num_reqs,
-                    max_context_len=self.model_config.context_len
-                    + extra_max_context_len,
-                    device=self.device,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                )
-        else:
-            # Draft worker shares req_to_token_pool with the target worker.
-            assert self.is_draft_worker
-
-        # Initialize token_to_kv_pool
-        is_dsa_model = is_deepseek_dsa(self.model_config.hf_config)
-        is_dsv4_model = is_deepseek_v4(self.model_config.hf_config)
-
-        self._validate_prefill_only_disable_kv_cache_pool_family(
-            is_dsa_model, is_dsv4_model, current_platform
-        )
-
-        if is_dsv4_model:
-            swa_page_size = self.page_size
-            if not _is_npu:
-                assert swa_page_size == 256, "In paged swa mode, page_size must be 256."
-
-            if self.is_draft_worker:
-                from sglang.srt.models.deepseek_v4_nextn import (
-                    COMPRESS_RATIO_NEXTN_LAYER,
-                )
-
-                compression_ratios = [
-                    COMPRESS_RATIO_NEXTN_LAYER
-                ] * self.num_effective_layers
-            else:
-                compression_ratios = self.model_config.compress_ratios
-
-            # NPU + DSV4 → paged-state subclass: the fused compressor kernel
-            # needs cache_mode=1 (paged); Atlas A3 rejects cache_mode=2 (ring),
-            # so the CUDA ring-buffer state path can't be shared. CUDA keeps
-            # DeepSeekV4TokenToKVPool unchanged; NPU recomputes state sizes below.
-            if _is_npu:
-                from sglang.srt.hardware_backend.npu.dsv4.dsv4_memory_pool import (
-                    DSV4NPUTokenToKVPool,
-                    npu_state_pool_size,
-                )
-
-                pool_cls = DSV4NPUTokenToKVPool
-                # Recompute state pool sizes for the NPU paged formula (CUDA's
-                # ring sizes are dropped here). Tail-only allocation keeps the
-                # per-req-budget formula sufficient at any prefill length: long
-                # prompts allocate only ``tail+128`` (c4) / ``tail`` (c128)
-                # slots (tail = seq_len % 128), and decode is drained by
-                # sliding eviction in ``ScheduleBatch._evict_swa``.
-                c4_state_pool_size = npu_state_pool_size(
-                    ratio=4,
-                    page_size=self.page_size,
-                    max_num_reqs=self.max_running_requests,
-                )
-                c128_state_pool_size = npu_state_pool_size(
-                    ratio=128,
-                    page_size=self.page_size,
-                    max_num_reqs=self.max_running_requests,
-                )
-            else:
-                pool_cls = DeepSeekV4TokenToKVPool
-                c4_state_pool_size = self.c4_state_pool_size
-                c128_state_pool_size = self.c128_state_pool_size
-
-            self.token_to_kv_pool = pool_cls(
-                max_num_reqs=self.max_running_requests,
-                # SWA ring is indexed by req_pool_idx; PD decode inflates req_to_token
-                # past max_running_requests (pre-alloc), so size to the real capacity.
-                num_req_slots=self.req_to_token_pool.req_to_token.shape[0],
-                swa_size=self.swa_max_total_num_tokens,
-                c4_size=self.c4_max_total_num_tokens,
-                c128_size=self.c128_max_total_num_tokens,
-                c4_state_pool_size=c4_state_pool_size,
-                c128_state_pool_size=c128_state_pool_size,
-                page_size=self.page_size,
-                swa_page_size=swa_page_size,
-                sliding_window=self.model_config.window_size,
-                dtype=self.kv_cache_dtype,
-                c4_state_dtype=self.c4_state_dtype,
-                c128_state_dtype=self.c128_state_dtype,
-                qk_nope_head_dim=self.model_config.qk_nope_head_dim,
-                qk_rope_head_dim=self.model_config.qk_rope_head_dim,
-                indexer_head_dim=self.model_config.index_head_dim,
-                layer_num=self.num_effective_layers,
-                device=self.device,
-                enable_memory_saver=self.server_args.enable_memory_saver,
-                compression_ratios=compression_ratios,
-                start_layer=self.start_layer,
-                end_layer=self.end_layer,
-                enable_hisparse=self.enable_hisparse,
-                online_mtp_max_draft_tokens=(
-                    self.server_args.max_speculative_num_draft_tokens or 0
-                ),
-            )
-        elif current_platform.is_out_of_tree() and not mambaish_config(
-            self.model_config
-        ):
-            if self.use_mla_backend and is_dsa_model:
-                PoolCls = current_platform.get_dsa_kv_pool_cls()
-                self.token_to_kv_pool = PoolCls(
-                    self.max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    kv_lora_rank=self.model_config.kv_lora_rank,
-                    qk_rope_head_dim=self.model_config.qk_rope_head_dim,
-                    layer_num=self.num_effective_layers,
-                    device=self.device,
-                    kv_cache_dim=calculate_mla_kv_cache_dim(
-                        model_config=self.model_config,
-                        kv_cache_dtype=self.kv_cache_dtype,
-                        server_args=self.server_args,
-                    ),
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    start_layer=self.start_layer,
-                    end_layer=self.end_layer,
-                    index_head_dim=get_dsa_index_head_dim(self.model_config.hf_config),
-                )
-            elif self.use_mla_backend:
-                PoolCls = current_platform.get_mla_kv_pool_cls()
-                self.token_to_kv_pool = PoolCls(
-                    self.max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    kv_lora_rank=self.model_config.kv_lora_rank,
-                    qk_rope_head_dim=self.model_config.qk_rope_head_dim,
-                    index_head_dim=(
-                        self.model_config.index_head_dim if is_dsa_model else None
-                    ),
-                    layer_num=self.num_effective_layers,
-                    device=self.device,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    start_layer=self.start_layer,
-                    end_layer=self.end_layer,
-                )
-            else:
-                PoolCls = current_platform.get_mha_kv_pool_cls()
-                self.token_to_kv_pool = PoolCls(
-                    self.max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    head_num=self.model_config.get_num_kv_heads(
-                        get_attention_tp_size()
-                    ),
-                    head_dim=self.model_config.head_dim,
-                    layer_num=self.num_effective_layers,
-                    device=self.device,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    start_layer=self.start_layer,
-                    end_layer=self.end_layer,
-                )
-        elif self.server_args.attention_backend == "ascend" and not mambaish_config(
-            self.model_config
-        ):
-            if self.is_hybrid_swa:
-                from sglang.srt.hardware_backend.npu.memory_pool_npu import (
-                    NPUMHATokenToKVPool,
-                )
-
-                kwargs = {}
-                if self.is_hybrid_swa_compress:
-                    kwargs = {
-                        "swa_head_num": max(
-                            1,
-                            self.model_config.hf_text_config.swa_num_key_value_heads
-                            // get_attention_tp_size(),
-                        ),
-                        "swa_head_dim": self.model_config.swa_head_dim,
-                        "swa_v_head_dim": self.model_config.swa_v_head_dim,
-                        "v_head_dim": self.model_config.v_head_dim,
-                    }
-                self.token_to_kv_pool = SWAKVPool(
-                    size=self.full_max_total_num_tokens,
-                    size_swa=self.swa_max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    head_num=self.model_config.get_num_kv_heads(
-                        get_attention_tp_size()
-                    ),
-                    head_dim=self.model_config.head_dim,
-                    swa_attention_layer_ids=self.model_config.swa_attention_layer_ids,
-                    full_attention_layer_ids=self.model_config.full_attention_layer_ids,
-                    enable_kvcache_transpose=False,
-                    device=self.device,
-                    token_to_kv_pool_class=NPUMHATokenToKVPool,
-                    **kwargs,
-                )
-            elif self.use_mla_backend:
-                from sglang.srt.hardware_backend.npu.memory_pool_npu import (
-                    NPUMLATokenToKVPool,
-                )
-
-                self.token_to_kv_pool = NPUMLATokenToKVPool(
-                    self.max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    kv_lora_rank=self.model_config.kv_lora_rank,
-                    qk_rope_head_dim=self.model_config.qk_rope_head_dim,
-                    index_head_dim=(
-                        self.model_config.index_head_dim if is_dsa_model else None
-                    ),
-                    layer_num=self.num_effective_layers,
-                    device=self.device,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    start_layer=self.start_layer,
-                    end_layer=self.end_layer,
-                )
-            else:
-                from sglang.srt.hardware_backend.npu.memory_pool_npu import (
-                    NPUMHATokenToKVPool,
-                )
-
-                self.token_to_kv_pool = NPUMHATokenToKVPool(
-                    self.max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    head_num=self.model_config.get_num_kv_heads(
-                        get_attention_tp_size()
-                    ),
-                    head_dim=self.model_config.head_dim,
-                    layer_num=self.num_effective_layers,
-                    device=self.device,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    start_layer=self.start_layer,
-                    end_layer=self.end_layer,
-                )
-        elif self.use_mla_backend and is_dsa_model:
-            PoolCls = (
-                HiSparseDSATokenToKVPool if self.enable_hisparse else DSATokenToKVPool
-            )
-            pool_kwargs = {}
-            if self.enable_hisparse:
-                from sglang.srt.mem_cache.sparsity import parse_hisparse_config
-
-                pool_kwargs["host_to_device_ratio"] = parse_hisparse_config(
-                    self.server_args
-                ).host_to_device_ratio
-            self.token_to_kv_pool = PoolCls(
-                self.max_total_num_tokens,
-                page_size=self.page_size,
-                dtype=self.kv_cache_dtype,
-                kv_lora_rank=self.model_config.kv_lora_rank,
-                qk_rope_head_dim=self.model_config.qk_rope_head_dim,
-                layer_num=self.num_effective_layers,
-                device=self.device,
-                kv_cache_dim=calculate_mla_kv_cache_dim(
-                    model_config=self.model_config,
-                    kv_cache_dtype=self.kv_cache_dtype,
-                    server_args=self.server_args,
-                ),
-                enable_memory_saver=self.server_args.enable_memory_saver,
-                start_layer=self.start_layer,
-                end_layer=self.end_layer,
-                index_head_dim=get_dsa_index_head_dim(self.model_config.hf_config),
-                **pool_kwargs,
-            )
-        elif self.use_mla_backend and not mambaish_config(self.model_config):
-            assert not is_dsa_model
-            if is_float4_e2m1fn_x2(self.kv_cache_dtype):
-                self.token_to_kv_pool = MLATokenToKVPoolFP4(
-                    self.max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    kv_lora_rank=self.model_config.kv_lora_rank,
-                    qk_rope_head_dim=self.model_config.qk_rope_head_dim,
-                    layer_num=self.num_effective_layers,
-                    device=self.device,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    start_layer=self.start_layer,
-                    end_layer=self.end_layer,
-                )
-            else:
-                self.token_to_kv_pool = MLATokenToKVPool(
-                    self.max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    kv_lora_rank=self.model_config.kv_lora_rank,
-                    qk_rope_head_dim=self.model_config.qk_rope_head_dim,
-                    layer_num=self.num_effective_layers,
-                    device=self.device,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    start_layer=self.start_layer,
-                    end_layer=self.end_layer,
-                )
-        else:
-            if self.is_hybrid_swa:
-                kwargs = {}
-                if self.is_hybrid_swa_compress:
-                    kwargs = {
-                        "swa_head_num": max(
-                            1,
-                            self.model_config.hf_text_config.swa_num_key_value_heads
-                            // get_attention_tp_size(),
-                        ),
-                        "swa_head_dim": self.model_config.swa_head_dim,
-                        "swa_v_head_dim": self.model_config.swa_v_head_dim,
-                        "v_head_dim": self.model_config.v_head_dim,
-                    }
-                self.token_to_kv_pool = SWAKVPool(
-                    size=self.full_max_total_num_tokens,
-                    size_swa=self.swa_max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    head_num=self.model_config.get_num_kv_heads(
-                        get_attention_tp_size()
-                    ),
-                    head_dim=self.model_config.head_dim,
-                    swa_attention_layer_ids=self.model_config.swa_attention_layer_ids,
-                    full_attention_layer_ids=self.model_config.full_attention_layer_ids,
-                    enable_kvcache_transpose=False,
-                    device=self.device,
-                    enable_kv_cache_copy=(
-                        self.server_args.speculative_algorithm is not None
-                    ),
-                    **kwargs,
-                )
-            elif config := mambaish_config(self.model_config):
-                extra_args = {}
-                if self.use_mla_backend:
-                    extra_args = {
-                        "kv_lora_rank": self.model_config.kv_lora_rank,
-                        "qk_rope_head_dim": self.model_config.qk_rope_head_dim,
-                    }
-                self.token_to_kv_pool = HybridLinearKVPool(
-                    page_size=self.page_size,
-                    size=self.max_total_num_tokens,
-                    dtype=self.kv_cache_dtype,
-                    head_num=self.model_config.get_num_kv_heads(
-                        get_attention_tp_size()
-                    ),
-                    head_dim=self.model_config.head_dim,
-                    # if draft worker, we only need 1 attention layer's kv pool
-                    full_attention_layer_ids=(
-                        [0]
-                        if self.is_draft_worker
-                        else [
-                            i
-                            for i in config.full_attention_layer_ids
-                            if self.start_layer <= i < self.end_layer
-                        ]
-                    ),
-                    enable_kvcache_transpose=False,
-                    device=self.device,
-                    mamba_pool=self.req_to_token_pool.mamba_pool,
-                    enable_memory_saver=self.server_args.enable_memory_saver,
-                    enable_kv_cache_copy=(
-                        self.server_args.speculative_algorithm is not None
-                    ),
-                    use_mla=self.use_mla_backend,
-                    start_layer=self.start_layer,
-                    **extra_args,
-                )
-            else:
-                if is_float4_e2m1fn_x2(self.kv_cache_dtype):
-                    self.token_to_kv_pool = MHATokenToKVPoolFP4(
-                        self.max_total_num_tokens,
-                        page_size=self.page_size,
-                        dtype=self.kv_cache_dtype,
-                        head_num=self.model_config.get_num_kv_heads(
-                            get_attention_tp_size()
-                        ),
-                        head_dim=self.model_config.head_dim,
-                        v_head_dim=self.model_config.v_head_dim,
-                        layer_num=self.num_effective_layers,
-                        device=self.device,
-                        enable_memory_saver=self.server_args.enable_memory_saver,
-                        start_layer=self.start_layer,
-                        end_layer=self.end_layer,
-                        enable_alt_stream=not self.server_args.enable_pdmux,
-                        enable_kv_cache_copy=(
-                            self.server_args.speculative_algorithm is not None
-                        ),
-                    )
-                else:
-                    pool_cls = (
-                        NoOpMHATokenToKVPool
-                        if self.server_args.prefill_only_disable_kv_cache
-                        else MHATokenToKVPool
-                    )
-                    self.token_to_kv_pool = pool_cls(
-                        self.max_total_num_tokens,
-                        page_size=self.page_size,
-                        dtype=self.kv_cache_dtype,
-                        head_num=self.model_config.get_num_kv_heads(
-                            get_attention_tp_size()
-                        ),
-                        head_dim=self.model_config.head_dim,
-                        v_head_dim=self.model_config.v_head_dim,
-                        layer_num=self.num_effective_layers,
-                        device=self.device,
-                        enable_memory_saver=self.server_args.enable_memory_saver,
-                        start_layer=self.start_layer,
-                        end_layer=self.end_layer,
-                        enable_alt_stream=not self.server_args.enable_pdmux,
-                        enable_kv_cache_copy=(
-                            self.server_args.speculative_algorithm is not None
-                        ),
-                    )
-
-        # Initialize token_to_kv_pool_allocator
-        need_sort = self.server_args.disaggregation_mode in ("decode", "prefill")
-        if self.token_to_kv_pool_allocator is None:
-            if current_platform.is_out_of_tree():
-                AllocatorCls = current_platform.get_paged_allocator_cls()
-                self.token_to_kv_pool_allocator = AllocatorCls(
-                    self.max_total_num_tokens,
-                    page_size=self.page_size,
-                    dtype=self.kv_cache_dtype,
-                    device=self.device,
-                    kvcache=self.token_to_kv_pool,
-                    need_sort=need_sort,
-                )
-            elif _is_npu and (
-                self.server_args.attention_backend == "ascend"
-                or is_dsv4_model
-                or hybrid_gdn_config(self.model_config) is not None
-            ):
-                if self.is_hybrid_swa:
-                    # DSV4 on NPU: SWA allocator subclass that also drives the
-                    # c4/c128 allocators, producing a DSV4OutCacheLoc per alloc.
-                    if is_dsv4_model:
-                        from sglang.srt.hardware_backend.npu.dsv4.dsv4_allocator import (
-                            DSV4NPUTokenToKVPoolAllocator,
-                        )
-
-                        swa_allocator_cls = DSV4NPUTokenToKVPoolAllocator
-                    else:
-                        swa_allocator_cls = SWATokenToKVPoolAllocator
-                    self.token_to_kv_pool_allocator = swa_allocator_cls(
-                        self.full_max_total_num_tokens,
-                        self.swa_max_total_num_tokens,
-                        page_size=self.page_size,
-                        dtype=self.kv_cache_dtype,
-                        device=self.device,
-                        kvcache=self.token_to_kv_pool,
-                        need_sort=need_sort,
-                    )
-                else:
-                    from sglang.srt.hardware_backend.npu.allocator_npu import (
-                        NPUPagedTokenToKVPoolAllocator,
-                    )
-
-                    self.token_to_kv_pool_allocator = NPUPagedTokenToKVPoolAllocator(
-                        self.max_total_num_tokens,
-                        page_size=self.page_size,
-                        dtype=self.kv_cache_dtype,
-                        device=self.device,
-                        kvcache=self.token_to_kv_pool,
-                        need_sort=need_sort,
-                    )
-            else:
-                if self.is_hybrid_swa:
-                    self.token_to_kv_pool_allocator = SWATokenToKVPoolAllocator(
-                        self.full_max_total_num_tokens,
-                        self.swa_max_total_num_tokens,
-                        page_size=self.page_size,
-                        dtype=self.kv_cache_dtype,
-                        device=self.device,
-                        kvcache=self.token_to_kv_pool,
-                        need_sort=need_sort,
-                    )
-                else:
-                    if self.enable_hisparse:
-                        from sglang.srt.mem_cache.sparsity import (
-                            parse_hisparse_config,
-                        )
-
-                        hisparse_cfg = parse_hisparse_config(self.server_args)
-                        self.token_to_kv_pool_allocator = (
-                            HiSparseTokenToKVPoolAllocator(
-                                self.max_total_num_tokens,
-                                page_size=self.page_size,
-                                dtype=self.kv_cache_dtype,
-                                device=self.device,
-                                kvcache=self.token_to_kv_pool,
-                                need_sort=need_sort,
-                                host_to_device_ratio=hisparse_cfg.host_to_device_ratio,
-                            )
-                        )
-                    elif self.page_size == 1:
-                        self.token_to_kv_pool_allocator = TokenToKVPoolAllocator(
-                            self.max_total_num_tokens,
-                            dtype=self.kv_cache_dtype,
-                            device=self.device,
-                            kvcache=self.token_to_kv_pool,
-                            need_sort=need_sort,
-                        )
-                    else:
-                        self.token_to_kv_pool_allocator = PagedTokenToKVPoolAllocator(
-                            self.max_total_num_tokens,
-                            page_size=self.page_size,
-                            dtype=self.kv_cache_dtype,
-                            device=self.device,
-                            kvcache=self.token_to_kv_pool,
-                            need_sort=need_sort,
-                        )
-
-            if self.enable_hisparse and is_dsv4_model:
-                assert self.is_hybrid_swa, "DeepSeek V4 HiSparse requires SWA mode."
-                self.token_to_kv_pool_allocator = (
-                    DeepSeekV4HiSparseTokenToKVPoolAllocator(
-                        self.token_to_kv_pool_allocator
-                    )
-                )
-
-            # DSV4-NPU: wire allocator back-ref into req_to_token_pool so its
-            # free(req) can release c4/c128 pool pages alongside the slot.
-            if hasattr(self.req_to_token_pool, "register_dsv4_allocator"):
-                self.req_to_token_pool.register_dsv4_allocator(
-                    self.token_to_kv_pool_allocator
-                )
-
-        else:
-            assert self.is_draft_worker
-            if self.is_hybrid_swa:
-                swa_allocator = getattr(
-                    self.token_to_kv_pool_allocator,
-                    "logical_attn_allocator",
-                    self.token_to_kv_pool_allocator,
-                )
-                assert swa_allocator.__class__ == SWATokenToKVPoolAllocator
-                self.token_to_kv_pool.full_to_swa_index_mapping = (
-                    swa_allocator.full_to_swa_index_mapping
-                )
-
-        # Defensive check: the explicit validation above should reject known
-        # unsupported pool families before allocation. Keep this guard here so
-        # future pool-selection refactors fail at boot instead of on first use.
-        if (
-            self.server_args.prefill_only_disable_kv_cache
-            and not self.is_draft_worker
-            and not isinstance(self.token_to_kv_pool, NoOpMHATokenToKVPool)
-        ):
-            raise RuntimeError(
-                "--prefill-only-disable-kv-cache expected NoOpMHATokenToKVPool but the "
-                f"runtime pool is {type(self.token_to_kv_pool).__name__}. This pool "
-                "family is not yet supported by --prefill-only-disable-kv-cache. "
-                "Supported configurations today: plain MHA models on CUDA with the FA "
-                "(fa3/fa4) prefill backend, --is-embedding, --chunked-prefill-size=-1, "
-                "--disable-radix-cache, no context-parallel attention, no HiSparse, "
-                "and --kv-cache-dtype != fp4_e2m1."
-            )
-
     def _apply_token_constraints(self: ModelRunner, token_capacity: int) -> int:
         return self.kv_cache_configurator._apply_token_constraints(token_capacity)
 
@@ -801,7 +96,24 @@ class ModelRunnerKVCacheMixin:
                 _get_dsv4_compress_state_dtypes()
             )
 
-        self._init_pools()
+        (
+            self.req_to_token_pool,
+            self.token_to_kv_pool,
+            self.token_to_kv_pool_allocator,
+        ) = self.kv_cache_configurator._init_pools(
+            max_total_num_tokens=self.max_total_num_tokens,
+            max_running_requests=self.max_running_requests,
+            full_max_total_num_tokens=getattr(self, "full_max_total_num_tokens", None),
+            swa_max_total_num_tokens=getattr(self, "swa_max_total_num_tokens", None),
+            c4_max_total_num_tokens=self.c4_max_total_num_tokens,
+            c128_max_total_num_tokens=self.c128_max_total_num_tokens,
+            c4_state_pool_size=self.c4_state_pool_size,
+            c128_state_pool_size=self.c128_state_pool_size,
+            c4_state_dtype=getattr(self, "c4_state_dtype", None),
+            c128_state_dtype=getattr(self, "c128_state_dtype", None),
+            req_to_token_pool=self.req_to_token_pool,
+            token_to_kv_pool_allocator=self.token_to_kv_pool_allocator,
+        )
 
     def _resolve_memory_pool_config(
         self: ModelRunner, pre_model_load_memory: int
